+++ dev_history.R
@@ -11,6 +11,7 @@ library(rhub)
 library(spelling)
 library(pkgdown)
 library(lintr)
+library(knitr)
 
 #check in different platforms
 rhub::validate_email()
@@ -55,7 +56,9 @@ usethis::use_github_action_check_standard()
 #https://www.rostrum.blog/2020/08/09/ghactions-pkgs/
 # usethis::use_pkgdown()
 # usethis::use_github_action("pkgdown")
+knitr::render(input = "README.Rmd")
 pkgdown::build_site()
+system("cp -avr README_files docs")
 
 #build documentation
 devtools::document()