+++ NEWS.md
@@ -1,4 +1,12 @@
+
+## Version 1.0.2 (24/2/2021)
+
+  + Simplified rf_spatial().
+  + Modified rf_tuning() to better tune models fitted with rf_spatial().
+  + Minor fixes in several other functions.
+
 ## Version 1.0.1 (22/2/2021)
-All 'sf' dependencies removed from the package.
+
+  + All 'sf' dependencies removed from the package.
 
 ## Version 1.0.0 is ready!